major update in warehouse, finance, shared modules

+++ Modules/Warehouse/Http/Controllers/Api/Document/DestroyWarehouseDocument.php
@@ -1,17 +1,18 @@
 <?php
 
-namespace Modules\Modules\Warehouse\Http\Controllers\Api\Document;
+namespace Modules\Warehouse\Http\Controllers\Api\Document;
 
-use Modules\Modules\Shared\Http\Controllers\BaseCrudHandler;
-use Modules\Modules\Shared\Services\Responder;
-use Modules\Modules\Warehouse\Models\WarehouseDocument;
+use Modules\Shared\Http\Controllers\BaseCrudHandler;
+use Modules\Shared\Services\Responder;
+use Modules\Warehouse\Models\WarehouseDocument;
 use OpenApi\Annotations as OA;
+use Symfony\Component\HttpFoundation\Response;
 
 /**
  * @OA\Delete(
  *     path="/api/warehouses/documents/{id}/delete",
  *     operationId="destroyWarehouseDocument",
- *     tags={"WarehouseDocuments"},
+ *     tags={"Warehouse > Documents"},
  *     summary="Delete existing warehouse document",
  *     description="Deletes a record and returns no content",
  *     @OA\Parameter(
@@ -35,15 +36,22 @@ use OpenApi\Annotations as OA;
  *     )
  * )
  */
-
 class DestroyWarehouseDocument extends BaseCrudHandler
 {
     public function execute(array $attributes = [])
     {
         $warehouseDocument = WarehouseDocument::whereId($attributes['id'])->firstOrFail();
+        if ($warehouseDocument->products()->count()) {
+            return Responder::error('سند مورد نظر دارای محصول است ابتدا آن ها را حذف کنید', null, Response::HTTP_UNPROCESSABLE_ENTITY);
+        }
 
-        $warehouseDocument->document->delete();
+        $warehouseDocument->documentable->delete();
 
         return Responder::success($warehouseDocument->delete());
+    }
+
+    public function authorize()
+    {
+        return true;
     }
 }
